cookieserver/requirements.txt
fastapi==0.95.1
gunicorn==20.1.0
pycryptodome==3.17
uvicorn==0.22.0
